[module view] move validation state to resource viewer
