+++ aptos-move/aptos-resource-viewer/Cargo.toml
@@ -14,10 +14,18 @@ rust-version = { workspace = true }
 
 [dependencies]
 anyhow = { workspace = true }
+aptos-logger = { workspace = true }
 aptos-types = { workspace = true }
 aptos-vm = { workspace = true }
 aptos-vm-environment = { workspace = true }
+aptos-vm-logging = { workspace = true }
 move-binary-format = { workspace = true }
 move-bytecode-utils = { workspace = true }
 move-core-types = { workspace = true }
 move-resource-viewer = { workspace = true }
+move-vm-runtime = { workspace = true }
+move-vm-types = { workspace = true }
+
+[dev-dependencies]
+aptos-types = { workspace = true, features = ["testing"] }
+move-vm-types = { workspace = true, features = ["testing"] }
