+++ vm-validator/Cargo.toml
@@ -15,16 +15,13 @@ rust-version = { workspace = true }
 [dependencies]
 anyhow = { workspace = true }
 aptos-logger = { workspace = true }
+aptos-resource-viewer = { workspace = true }
 aptos-storage-interface = { workspace = true }
 aptos-types = { workspace = true }
 aptos-vm = { workspace = true }
-aptos-vm-environment = { workspace = true }
-aptos-vm-logging = { workspace = true }
 fail = { workspace = true }
-move-binary-format = { workspace = true }
 move-core-types = { workspace = true }
 move-vm-runtime = { workspace = true }
-move-vm-types = { workspace = true }
 rand = { workspace = true }
 
 [dev-dependencies]
@@ -32,7 +29,7 @@ aptos-cached-packages = { workspace = true }
 aptos-crypto = { workspace = true }
 aptos-db = { workspace = true }
 aptos-executor-test-helpers = { workspace = true }
-aptos-gas-schedule = { workspace = true, features = ["testing"] } 
+aptos-gas-schedule = { workspace = true, features = ["testing"] }
 aptos-temppath = { workspace = true }
 aptos-types = { workspace = true, features = ["testing"] }
 aptos-vm-genesis = { workspace = true }
